refactor(Cars.java, RacingGame.java): car리스트를 일급 컬렉션으로 분리
- car리스트를 일급 컬렉션으로 분리
