+++ src/main/java/racingcar/models/Cars.java
@@ -1,0 +1,53 @@
+package racingcar.models;
+
+import racingcar.utils.RandomNumber;
+
+import java.util.Comparator;
+import java.util.List;
+import java.util.stream.Collectors;
+
+public class Cars {
+
+    private final List<Car> cars;
+    private final RandomNumber randomNumber;
+
+    public Cars(List<Car> cars) {
+        this.randomNumber = RandomNumber.getInstance();
+        this.cars = cars;
+    }
+
+    public void startOneTurn() {
+        cars.forEach(car -> {
+            int randomNumber = this.randomNumber.getRandomNumber();
+            car.goForward(randomNumber);
+        });
+    }
+
+    public String findWinner() {
+        final int farthestPosition = findFarthestPosition();
+        return getWinners(farthestPosition);
+    }
+
+    private int findFarthestPosition() {
+        final List<Integer> sortedCars = cars.stream()
+                .sorted(Comparator.comparing(Car::getPosition).reversed())
+                .map(Car::getPosition)
+                .collect(Collectors.toList());
+
+        if (!sortedCars.isEmpty()) {
+            return sortedCars.get(0);
+        }
+        return 0;
+    }
+
+    private String getWinners(final int farthestPosition) {
+        return cars.stream()
+                .filter((car) -> car.isWinner(farthestPosition))
+                .map(Car::getName)
+                .collect(Collectors.joining(", "));
+    }
+
+    public List<Car> getCars() {
+        return cars;
+    }
+}
